requirements.txt
dj-config-url==0.1.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9
django-bootstrap3==6.2.2
django-debug-toolbar==1.4
django-extensions==1.5.9
django-toolbelt==0.0.1
djangorestframework==3.3.1
gunicorn==19.4.1
psycopg2==2.6.1
six==1.10.0
sqlparse==0.1.18
static3==0.6.1
wheel==0.24.0
